No reason CardHolder shouldn't be a class
The original design of CardHolder was as a module that could be included
in any class (i.e. a Dealer, a Player, etc.), but after a few
refactorings it became obvious that the various classes I had in mind
didn't have any class-specific logic that couldn't be defined just as
well as a subclass.